+++ rqes-data-classes/src/commonMain/kotlin/at/asitplus/rqes/CredentialListRequest.kt
@@ -4,9 +4,14 @@ import at.asitplus.rqes.enums.CertificateOptions
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
+/**
+ * Data class that implements credentialInfo defined
+ * in CSC-API v2.0.0.2 Ch. 11.4 "credentials/list" and Ch. 11.5 "credentials/info"
+ */
 @Serializable
-data class CscCredentialListRequest(
+data class CredentialListRequest(
     /**
+     * CSC REQUIRED-CONDITIONAL.
      * The identifier associated to the identity of the credential owner. This parameter
      * SHALL NOT be present if the service authorization is user-specific (see NOTE below).
      * In that case the userID is already implicit in the service access token passed in the
@@ -16,16 +21,18 @@ data class CscCredentialListRequest(
     val userID: String? = null,
 
     /**
+     * CSC OPTIONAL.
      * Request to return the main information included in the public key certificate and
      * the public key certificate itself or the certificate chain associated to the credentials.
-     * The default value is “false”
+     * The default value is [Boolean.false]
      */
     @SerialName("credentialInfo")
     val credentialInfo: Boolean? = false,
 
     /**
+     * CSC OPTIONAL-CONDITIONAL.
      * Specifies which certificates from the certificate chain SHALL be returned
-     * The default value is “single”
+     * The default value is [CertificateOptions.SINGLE]
      */
     @SerialName("certificates")
     val certificates: CertificateOptions? = CertificateOptions.SINGLE,
@@ -33,7 +40,7 @@ data class CscCredentialListRequest(
     /**
      * Request to return various parameters containing information from the end entity
      * certificate(s).
-     * The default value is “false”
+     * The default value is [Boolean.false]
      */
     @SerialName("certInfo")
     val certInfo: Boolean? = false,