+++ rqes-data-classes/src/commonMain/kotlin/at/asitplus/rqes/CredentialListResponse.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class CscCredentialListResponse(
+data class CredentialListResponse(
     /**
      * One or more credentialID(s) associated with the provided or implicit userID.
      */