+++ rqes-data-classes/src/commonMain/kotlin/at/asitplus/rqes/collection_entries/DocumentDigest.kt
@@ -23,7 +23,7 @@ import kotlinx.serialization.json.JsonObject
  * CSC: Class used as part of [SignatureRequestParameters]
  */
 @Serializable
-data class CscDocumentDigest(
+data class DocumentDigest(
     /**
      *
      * One or more hash values representing one or more SDRs. This
@@ -95,7 +95,7 @@ data class CscDocumentDigest(
         if (this === other) return true
         if (other == null || this::class != other::class) return false
 
-        other as CscDocumentDigest
+        other as DocumentDigest
 
         if (!hashes.contentEquals(other.hashes)) return false
         if (hashAlgorithmOid != other.hashAlgorithmOid) return false