+++ rqes-data-classes/src/commonMain/kotlin/at/asitplus/rqes/collection_entries/KeyParameters.kt
@@ -1,11 +1,12 @@
 package at.asitplus.rqes.collection_entries
 
+import at.asitplus.signum.indispensable.X509SignatureAlgorithm
 import at.asitplus.signum.indispensable.asn1.ObjectIdentifier
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class CscKeyParameters(
+data class KeyParameters(
     /**
      * REQUIRED.
      * The status of the signing key of the credential:
@@ -18,7 +19,7 @@ data class CscKeyParameters(
      * The list of OIDs of the supported key algorithms
      */
     @SerialName("algo")
-    val algo: Collection<ObjectIdentifier>,
+    val algo: Set<ObjectIdentifier>,
 
     /**
      * REQUIRED.
@@ -35,6 +36,18 @@ data class CscKeyParameters(
     @SerialName("curve")
     val curve: ObjectIdentifier? = null,
 ) {
+    init {
+        require(
+            algo.intersect(
+                listOf(
+                    X509SignatureAlgorithm.ES256.oid,
+                    X509SignatureAlgorithm.ES384.oid,
+                    X509SignatureAlgorithm.ES512.oid
+                ).toSet()
+            ) != emptySet<ObjectIdentifier>() || curve == null
+        )
+    }
+
     enum class KeyStatusOptions {
         /**
          * the signing key is enabled and can be used for signing.