+++ rqes-data-classes/src/commonMain/kotlin/at/asitplus/rqes/serializers/CscSignatureRequestParameterSerializer.kt
@@ -1,6 +1,6 @@
 package at.asitplus.rqes.serializers
 
-import at.asitplus.rqes.CscSignatureRequestParameters
+import at.asitplus.rqes.QtspSignatureRequest
 import at.asitplus.rqes.SignDocParameters
 import at.asitplus.rqes.SignHashParameters
 import kotlinx.serialization.json.JsonContentPolymorphicSerializer
@@ -8,7 +8,7 @@ import kotlinx.serialization.json.JsonElement
 import kotlinx.serialization.json.jsonObject
 
 object CscSignatureRequestParameterSerializer :
-    JsonContentPolymorphicSerializer<CscSignatureRequestParameters>(CscSignatureRequestParameters::class) {
+    JsonContentPolymorphicSerializer<QtspSignatureRequest>(QtspSignatureRequest::class) {
     override fun selectDeserializer(element: JsonElement) = when {
         "hashes" in element.jsonObject -> SignHashParameters.serializer()
         else -> SignDocParameters.serializer()